refactor: update folder dir
