+++ app/layout.tsx
@@ -1,6 +1,6 @@
 import "@/app/styles/globals.css";
 import { MainNav } from "@/components/main-nav";
-import { constructMetadata } from "@/lib/utils/functions";
+import { constructMetadata } from "@/lib/utils";
 import { crimsonPro, geistMono, inter, satoshi } from "@/styles/fonts";
 import { Viewport } from "next";
