+++ tsconfig.json
@@ -22,7 +22,7 @@
     "paths": {
       "@/app/*": ["./app/*"],
       "@/components/*": ["components/*"],
-      "@/lib/*": ["app/lib/*"],
+      "@/lib/*": ["lib/*"],
       "@/styles/*": ["app/styles/*"]
     }
   },
